Added partitioned I/O support.  Made string functionality same as QDP. Standardized fixed numeric types.  No support for parallel I/O or file appending yet, but included placeholders.  -CD

+++ include/lime_fixed_types.h
@@ -1,9 +1,11 @@
 #ifndef LIME_FIXED_TYPES_H
 #define LIME_FIXED_TYPES_H
 
+#include <stdint.h>
+
 /* Should be CONFIG'd */
-typedef unsigned long long  n_uint64_t;
-typedef unsigned long       n_uint32_t;
-typedef unsigned short      n_uint16_t;
+typedef uint64_t  n_uint64_t;
+typedef uint32_t  n_uint32_t;
+typedef uint16_t  n_uint16_t;
 
 #endif
